# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ghost Coach API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activity templates
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityTemplate"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create activity template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityTemplateInput"
      responses:
        "201":
          description: Created activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityTemplate"
  /activities/{id}:
    put:
      operationId: updateActivity
      tags: [activities]
      summary: Update activity template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityTemplateInput"
      responses:
        "200":
          description: Updated activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityTemplate"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete activity template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List saved locations
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Create saved location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Created location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /locations/{id}:
    put:
      operationId: updateLocation
      tags: [locations]
      summary: Update saved location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "200":
          description: Updated location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /locations/current:
    post:
      operationId: setCurrentLocation
      tags: [locations]
      summary: Set current active location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CurrentLocationInput"
      responses:
        "200":
          description: Location set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
...
components:
  schemas:
...
    ActivityTemplate:
      type: object
      required: [id, userId, name, category, goalTags, minimumMinutes, preferredMinutes, maximumMinutes, intensity, requiresFloorSpace, requiresPool, requiresStairs, requiresShower, requiresEquipment, minimumPrivacy, publicSuitability, weatherMode, allowedLocationTypes, preferenceScore, active, isPhysioApproved, version, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        category:
          type: string
        goalTags:
          type: array
          items:
            type: string
        minimumMinutes:
          type: integer
        preferredMinutes:
          type: integer
        maximumMinutes:
          type: integer
        intensity:
          type: string
        requiresFloorSpace:
          type: boolean
        requiresPool:
          type: boolean
        requiresStairs:
          type: boolean
        requiresShower:
          type: boolean
        requiresEquipment:
          type: array
          items:
            type: string
        minimumPrivacy:
          type: string
        publicSuitability:
          type: string
        weatherMode:
          type: string
        allowedLocationTypes:
          type: array
          items:
            type: string
        preferenceScore:
          type: integer
        active:
          type: boolean
        isPhysioApproved:
          type: boolean
        instructionsMarkdown:
          type: ["string", "null"]
        version:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    ActivityTemplateInput:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
        category:
          type: string
        goalTags:
          type: array
          items:
            type: string
        minimumMinutes:
          type: integer
        preferredMinutes:
          type: integer
        maximumMinutes:
          type: integer
        intensity:
          type: string
        requiresFloorSpace:
          type: boolean
        requiresPool:
          type: boolean
        requiresStairs:
          type: boolean
        requiresShower:
          type: boolean
        requiresEquipment:
          type: array
          items:
            type: string
        minimumPrivacy:
          type: string
        publicSuitability:
          type: string
        weatherMode:
          type: string
        allowedLocationTypes:
          type: array
          items:
            type: string
        preferenceScore:
          type: integer
        active:
          type: boolean
        isPhysioApproved:
          type: boolean
        instructionsMarkdown:
          type: string
    Location:
      type: object
      required: [id, userId, label, timezone, locationType, hasFloorSpace, hasPool, hasStairs, hasIndoorWalk, hasShower, publicPrivacyLevel, typicalTravelOverheadMinutes, isDefault, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        label:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        timezone:
          type: string
        locationType:
          type: string
        hasFloorSpace:
          type: boolean
        hasPool:
          type: boolean
        hasStairs:
          type: boolean
        hasIndoorWalk:
          type: boolean
        hasShower:
          type: boolean
        publicPrivacyLevel:
          type: string
        typicalTravelOverheadMinutes:
          type: integer
        isDefault:
          type: boolean
        lastConfirmedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    LocationInput:
      type: object
      required: [label, locationType]
      properties:
        label:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        timezone:
          type: string
        locationType:
          type: string
        hasFloorSpace:
          type: boolean
        hasPool:
          type: boolean
        hasStairs:
          type: boolean
        hasIndoorWalk:
          type: boolean
        hasShower:
          type: boolean
        publicPrivacyLevel:
          type: string
        typicalTravelOverheadMinutes:
          type: integer
        isDefault:
          type: boolean
    CurrentLocationInput:
      type: object
      required: [locationId]
      properties:
        locationId:
          type: string
...